requirements.txt
PyYAML==6.0
fastapi==0.95.0
uvicorn==0.21.1
httpx==0.23.3
pytest==7.2.2
